+++ src/auth/auth-token-error-code.ts
@@ -1,0 +1,6 @@
+export enum BBAuthTokenErrorCode {
+  Unspecified = 0,
+  NotLoggedIn,
+  InvalidEnvironment,
+  PermissionScopeNoEnvironment
+}